requirements.txt
aiogram>=3.4.1,<4.0.0
yt-dlp>=2024.03.10
python-dotenv>=1.0.1

